pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task Manager is a simple web application that helps users manage their tasks."
authors = ["vitallcore <vitaslipknotgmail.ru>"]
readme = "README.md"
repository = "https://github.com/vitallcore/python-project-52.git"
classifiers = [
    "Framework :: Django :: 5.1",
    "Programming Language :: Python :: 3.10"
]
packages = [
    { include = "task_manager" },
]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
python-dotenv = "^1.0.1"
dj-database-url = "^2.1.0"
gunicorn = "^22.0.0"
uvicorn = "^0.29.0"
psycopg2-binary = "^2.9.9"
django-bootstrap5 = "^24.2"
whitenoise = "^6.6.0"
django-filter = "^24.2"
rollbar = "0.16.3"
djangorestframework = "^3.15.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
ipython = "^8.24.0"
coverage = "^7.5.3"
ruff = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
